Windows build error.

+++ Engine/Containers/Freelist.cpp
@@ -4,7 +4,7 @@
 #include "Core/EngineLogger.hpp"
 #include "Platform/Platform.hpp"
 
-void Freelist::Create(unsigned long total_size) {
+void Freelist::Create(size_t total_size) {
 	// Enough space to hold state, plus array for all nodes.
 	MaxEntries = (total_size / sizeof(void*));	// NOTO: This might have a remainder, but ok.
 	TotalSize = total_size;
@@ -40,7 +40,7 @@ void Freelist::Destroy() {
 	}
 }
 
-bool Freelist::AllocateBlock(unsigned long size, unsigned long * offset) {
+bool Freelist::AllocateBlock(size_t size, size_t* offset) {
 	if (offset == nullptr || ListMemory == nullptr) {
 		return false;
 	}
@@ -81,7 +81,7 @@ bool Freelist::AllocateBlock(unsigned long size, unsigned long * offset) {
 	return false;
 }
 
-bool Freelist::FreeBlock(unsigned long size, unsigned long long offset) {
+bool Freelist::FreeBlock(size_t size, size_t offset) {
 	if (ListMemory == nullptr || size == 0) {
 		return false;
 	}
@@ -160,7 +160,7 @@ bool Freelist::FreeBlock(unsigned long size, unsigned long long offset) {
 	return false;
 }
 
-bool Freelist::Resize(unsigned long new_size) {
+bool Freelist::Resize(size_t new_size) {
 	if (ListMemory == nullptr || new_size < TotalSize) {
 		return false;
 	}
@@ -253,7 +253,7 @@ void Freelist::Clear() {
 	}
 }
 
-unsigned long Freelist::GetFreeSpace() {
+size_t Freelist::GetFreeSpace() {
 	if (ListMemory == nullptr) {
 		return 0;
 	}
